refactor home page
