+++ src/app/page.tsx
@@ -2,86 +2,15 @@
 import { useMutation, useQuery } from "convex/react";
 import { useState } from "react";
 import { api } from "../../convex/_generated/api";
-import { Hero } from "../../components/Hero";
 import headerImage from "../../public/images/kurt-and-cecile-proposal.jpg";
-import flowerImage from "../../public/images/flower.png";
-import ovalImage from "../../public/images/cecile-kurt-wedding-oval.png";
-import Image from "next/image";
+import { Hero } from "@/components/Hero";
+import { OurStory } from "@/components/OurStory";
 
 export default function Home() {
   return (
     <main className="overflow-hidden font-bitter">
-      {/* <HeroImage /> */}
-      <div className="fixed  w-full h-screen -z-10 flex justify-center">
-        <Hero />
-        <header className="flex flex-col z-10 content-center justify-center">
-          <h2 className="uppercase text-xs tracking-widest text-gray-300 self-center">
-            we are getting married
-          </h2>
-          <h1 className="tracking-tight text-gray-200 my-10 md:text-9xl lg:text-9xl font-brush">
-            Kurt <strong className="font-medium mx-2">&</strong> Cecile
-          </h1>
-          <div className="flex justify-center content-center">
-            <div className="bg-orange-200/35 w-full h-[1px] self-center" />
-            <h3 className="uppercase text-gray-300 w-full text-center tracking-wide">
-              Save the Date
-            </h3>
-            <div className="bg-orange-200/35 w-full h-[1px] self-center" />
-          </div>
-          <p className="self-center mt-10 font-extralight text-gray-300 text-3xl py-4 px-8">
-            August 28th 2025
-          </p>
-          <a
-            href="google.com"
-            target="_blank"
-            className="self-center mt-10 font-thin text-gray-300 cursor-pointer text-3xl border-2 border-orange-200/35 py-4 px-8 z-50"
-          >
-            See Our Registry
-          </a>
-        </header>
-      </div>
-
-      <section className="h-[100vw] translate-y-[100vh] bg-gray-100">
-        <div className="w-full pt-28 flex justify-center flex-col">
-          <Image
-            alt="wedding flower"
-            className="w-24 h-11 self-center"
-            src={flowerImage}
-          />
-          <h4 className="text-gray-800 text-5xl font-light text-center">
-            Our Story
-          </h4>
-          <p className="max-w-[1000px] text-gray-500 text-center font-light py-4 self-center tracking-wide">
-            Lorem ipsum, dolor sit amet consectetur adipisicing elit. Pariatur
-            repellendus totam modi voluptas cum commodi aspernatur ut minima a,
-            excepturi odit vel, fuga, suscipit dolores possimus quam
-            reprehenderit voluptates dolor. Perferendis totam eius aut molestias
-          </p>
-          <div className="mt-10 self-center w-full flex space-x-10 px-7 max-w-[1300px]">
-            <div>
-              <h5 className="text-gray-800 text-4xl mb-7">Cecile Badenhorst</h5>
-              <p className="text-gray-500 font-light text-sm leading-6 tracking-wide text-right">
-                Lorem ipsum, dolor sit amet consectetur adipisicing elit.
-                Cumque, doloremque earum fugit voluptatem repudiandae unde
-                necessitatibus. Recusandae, itaque praesentium voluptatibus modi
-                veritatis reiciendis ea ducimus, consectetur delectus, excepturi
-                omnis eaque.
-              </p>
-            </div>
-            <Image alt="Cecile and kurt oval photo" src={ovalImage} />
-            <div className="mt-32">
-              <h5 className="text-gray-800 text-4xl mb-7">Kurt Ruether</h5>
-              <p className="text-gray-500 font-light text-sm leading-6 tracking-wide">
-                Lorem ipsum, dolor sit amet consectetur adipisicing elit.
-                Cumque, doloremque earum fugit voluptatem repudiandae unde
-                necessitatibus. Recusandae, itaque praesentium voluptatibus modi
-                veritatis reiciendis ea ducimus, consectetur delectus, excepturi
-                omnis eaque.
-              </p>
-            </div>
-          </div>
-        </div>
-      </section>
+      <Hero />
+      <OurStory />
     </main>
   );
 }
